Okay!
We've figured out what caused the math error in the interval changes based on humidity and temperature.

It's the fact that the interval is on the right side of the equation in the humidity interval. You're going to have to adjust the algorithm with a constand for that decay rather than take into account what that interval is OR do some pretty advanced math.